feat: add a totalCount that returns a current condition

+++ backend/analyze-service/src/main/java/com/nine/project/analyze/dto/resp/RankRespDTO.java
@@ -29,5 +29,5 @@ public class RankRespDTO {
     /**
      * 总数
      */
-    private Long totalCount;
+    private Integer totalCount;
 }
